requirements.txt
fastapi==0.97.0
numpy==1.24.3
opencv_python==4.7.0.72
opencv_python_headless==4.7.0.72
Pillow==9.5.0
Pillow==9.5.0
pydantic==1.10.9
rembg==2.0.43
starlette==0.28.0
torch==2.0.1
torchvision==0.15.2
uvicorn==0.22.0
Werkzeug==2.3.6
